Extract socketListener.
